run: Add per test check for Suricata version
Some tests in the latest versions might fail due to compatibility issues
because of the changes in features with every version. In order to avoid
failure in such cases, add per test checks that allow to check for the
current Suricata version and in case it does not match the version of
Suricata required to run that particular test, skip that test with an
appropriate message.

Sample conf
-----------
```
requires:
  features:
    - HAVE_LIBJANSSON

checks:

  # Check that we only have one alert event type in eve.
  - filter:
      version: 4.0.2
      count: 1
      match:
        event_type: alert

  # Check how many lines were logged to fast.log.
  - shell:
      args: cat fast.log | wc -l | xargs
      expect: 1
```

Running `tests/alert-testmyids` with above configuration (test.yaml)
gives the following output when run from a directory having Suricata
version 5.0.0-dev.
```
===> alert-testmyids: SKIPPED: Suricata v4.0.2 not found
```

In case someone specifies `min-version` as well as `version` as a part
of the configuration, the test fails with an error as below.
```
===> alert-testmyids: FAIL: Specify either min-version or version
```

Closes redmine ticket #2924