Give test count as output not check count
With the current setup, after running suricata-verify, the output would
give stats about the passed, failed or skipped checks. However, it is
cleaner to get the total stats about tests only.

The priority order for the output is:
FAILED
SKIPPED
PASSED

i.e. if a check or sub-test has failed, the entire test will be counted
as failed in the final output. Similarly, if a check has been skipped,
the entire test will be counted as skipped in the final output.

Closes redmine ticket #3172.