github-ci: test against supported Suricata branches
The idea is not to test every compile time option, but a somewhat
standard once over of our supported branches to make sure tests
that are version specific are flagged as such.

Will also allow us to quickly spot PRs that fail that may
be depending on a suricata fix/pr.