tests: add test for max alert default & noalerts
Since the default max alert value is 15 and we want to ensure that that
value is enforced if the configuration is set with something invalid,
this test has packet-max-alert set to 0 (which will fall back to 15,
the default). We have 15 'noalert' rules and 1 alert rule that gets
triggered.

Related to task #4942