tests/etopen Update for dsize validation
This commit creates a test for Suricata versions 7 and later with the
updated dsize/content validation logic.

Note that the logic detects one rule that is written incorrectly. It has
been reported to Proofpoint. Until the ruleset has been adjusted, the
test rules that check results expect a single failure (due to the new
dsize validation)