tests: add test for unseen http midstream traffic
In a pcap where just `http` midstream traffic is seen, Suri is
unable to see the packtes as `http` traffic (Wireshark tags them
correctly).

This also seems to result in Suri sometimes not adding the packet
payload to the associated alert event in the eve-log.

`bug-5437-01` has the pcap where http packets are not seen `bug-5437-02`
has a more complete pcap, and the same packets are properly
identified by Suri.

Related to
Bug #5437