tests: add test for unseen http midstream traffic
In a pcap where just a `http` midstream traffic packets are seen, Suri is
unable to see those as `http` packets (Wireshark tags them correctly).

This also seems to result in Suri sometimes not adding the packet
payload to the associated alert event in the eve-log.

`unseen-http-stream-01` has the pcap where http packets are not seen
`unseen-http-stream-02` has a more complete pcap, where the same packets
are properly identified by Suri

Related to
Bug #5437