eve-validation: always build and run eve-validator
Always build and run the eve-validator as part of Suricata-Verify. This
should only add extra time on first run, or after changes to the
"eve-validator" or an upgrade in the Rust toolchain.

Validation can be disabled with the --no-validation command line option.

Validation will not be run if the Suricata source tree does not contain
etc/schema.json.

Validation will also not run on Windows as there is an issue calling the
validation command that needs further looking into.