runner: error out on errors from the runner
Currently exceptions from tests that are not explicitly handled are lost
due to the way Python's multiprocessing module works.  This means that
programming errors in the runner are silently ignore with the test not
being run or counted.

Instead, log the traceback for any unexpected exception and terminate
the runner, as this is programming error in the runner and should always
be fatal.