runner: handle binary output from suricata stderr/stdout
Don't attempt to decode output from Suricata stderr/stdout as utf-8, it
is required for relaying the output from Suricata to the log files,
in fact, its not even desired.  The log files should have a verbatim
copy of the output for analysis.

Only attempt to utf-8 decode the output when logging in verbose mode,
and then if that fails, fallback to logging the data as a byte buffer.