tests: fix bad http host rule tests
The test.yaml files were missing the command set to compare eve.json
output and to run without a pcap file, therefore being simply skipped
for lack of a pcap file.

Also took the opportunity to make these compatible with new error
message formats for Suricata 7.

Test 1 also had a typo in the expected message to be checked, making it
fail.