tests: test defrag exception policy with drop-flow
Defrag memcap and flow memcap do not support flow action for the
exception policies, as there is no flow when the exception condition is
hit. In such cases, the exception policy must be considered as
`drop-packet`. This commit changes the defrag exception policy test to
check if this behavior is working.