tests: showcase flow.action lack of update
It seems that in certain cases as seen in this test, flow.action isn't
updated, even if, say, all packets from the flow are dropped.

Maybe this is due to the rule not being applied directly to the flow,
but to each packet individually. But considering we are using a flow
keyword, it seems that the engine should pass over the drop action to
flow.action, at least in the flow event.

Bug #6976