tests: update filestore test
The signature was about storing on all events of the flow. After
fixing Suricata, this is really doing that. As a result the test
is failing.

By updating the signature to store on the tx only, we are now
testing what was looking like the purpose of the test which is
to only extract the file corresponding to the URL.