tests/lua: Test for issue 7466
This test was contributed by the reporter of issue 7466. Slight
modifications not affecting the structure nor results were made.

This test exercises the problem reported in the issue and a memory leak
will be reported on Suricata binaries without the fix (ASAN-only).

Non-ASAN builds will not report the issue (regardless of whether the
Suricata fix is included).