tests: remove exit-code check for datasets-absolute-allowed on v8
With the commit in Suricata to skip adding localstatedir when a full
path is provided, the S-V test does not exit with 1 anymore but rather
with 0 since it succeeds.

This commits updates the previous test to run with Suricata prior to
verison 8 and a dedicated copy of the test to work with Suricata 8
without the need to check the exit code anymore, since it's 0.

The new test is split between Unix like systems and Windows by using
different absolute paths for the filenames.