runner: convert to multi-threading from multi-processing
Multi-processing has issues on Windows, moving to threading does allow
it to work on Windows (with a few issues to still figure out).

This removes the single thread runner, instead for Windows and Mac
we'll just use one job for now, and change that default as we prove
its reliaable.

Update ctrl-c handling as well, for reliable ctrl-c cancellation.