workflows: only run the modified tests
Running all the tests on PR or push event of suricata-verify is
irrelevant as the tests are supposed to be exclusive of one another.
Make sure that unless there's a framework or workflows change, only the
tests that are modified or added are run. This saves CI resources that
would otherwise be unnecessarily spent.

This does not affect the testing and coverage of the Suricata codebase
as when there's an s-v PR, Suricata's workflow is to clone the entire
repo and run all the tests in there.